Update FastaToKmers to work on a single epitope length
